Temporarily switch to prebuilt gnss libraries
bluejay is based on android-12.1.0_r13 AOSP tag which is on older
codebase than android-12.1.0_r11 which we are currently using.

r11 contains incompatible blobs in device/google/gs101 that are
used for oriole/raven devices. Switch to prebuilt blobs to be safe
and make the least amount of changes in device/google/gs101.

Test: gps works on bluejay compiled on android-12.1.0_r11
Change-Id: I4a822d527710f181012d7fb1f4d24bfba7e76096